dynamic_llm_requirements.txt
# Core dependencies
simplerllm>=0.1.0  # Adjust version as needed
pydantic>=1.8.0,<2.0.0  # Version compatible with SimplerLLM
python-dotenv>=0.19.0

# FastAPI dependencies (for API server)
fastapi>=0.68.0
uvicorn>=0.15.0

# Testing dependencies
pytest>=6.2.5
pytest-asyncio>=0.15.1
